techstore/pyproject.toml
[tool.poetry]
name = "techstore"
version = "0.1.0"
description = "Sistema SaaS para gestión de comercios técnicos"
authors = ["lucast086 <turlettilucasdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
django-tenants = "^3.5.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
coverage = "^7.4.1"
black = "^24.1.1"
isort = "^5.13.2"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
